Rename msal.py to custom_browser_credential.py
