+++ msgraph/cli/core/custom_browser_credential.py
